Visual Studio: Use precompiled header, /MP

+++ src/kpnodes.h
@@ -61,7 +61,7 @@ public:
     template<typename F>
     void Iterate(F &f)
     {
-        for (int index = 0; index < ids.size(); ++index)
+        for (size_t index = 0; index < ids.size(); ++index)
         {
             f.Iterate(GetNodeFor(ids[index]));
         }
